Disabled Spread and Force-directed layout pre-diffusion
Re-drawing the full graph with these layouts is expensive and
lags the browser. It also looks like a hairball. The option to select
these layouts is now disabled on start up (when the full network is
loaded in the view).

The choice to select these layouts is enabled when the diffusion
experiment is complete, and the result nodes are shown.