Labels are now colored according to their rank in LOO experiments
By default, I had labels colored red. Now, they are colored the
same as other proteins during cross-validation experiments
(ie the color now corresponds to how well the label is connected
to the rest of the label set).

This may be a bit counter-intuitive to novice users. Not sure
if this was the right thing to do. I may revert this change.