Fixed infinite specificity score
Specificity scores are calculated for all terms in the ontology.

If the annotation corpus has 0 annotations for a particular GO term,
 that  GO term would recieve specificty of "np.Infinity" (a result of
a 0 divided by 0 operation).

I inserted a line that reassigns final "np.Infinity" scores to 0.

This is a band-aid fix. Ideally, I should filter the ontology tree
to only include terms that appear in the passed annotation corpus.
That would preclude division by zero issue.